Message input needs to be fixed. not display and to much xml. refactor
