+++ client/components/wireframes/GroupsScreen/Group/Header.tsx
@@ -7,13 +7,13 @@ interface IGroupHeaderProps {
 
 function Header({ name, moderator }: IGroupHeaderProps) {
   return (
-    <>
-      <h2 className=" text-left text-xl pb-2">{name} Group</h2>
-      <h3 className="text-left text-base pb-1">Moderator: {moderator}</h3>
-      <h3 className="text-left text-xs pb-1">
+    <div className="flex flex-col content-center">
+      <h2 className="text-xl pb-2">{name} Group</h2>
+      <h3 className="text-base pb-1">Moderator: {moderator}</h3>
+      <h3 className="text-xs pb-1">
         Deadline For Weekly Drawling: 10/21/2023 5:30PM
       </h3>
-    </>
+    </div>
   )
 }
