+++ client/components/wireframes/GroupsScreen/Group/LogDisplay/Message/Icon.tsx
