+++ client/components/wireframes/GroupsScreen/Group/LogDisplay/Message/SelectedByModal.tsx
