+++ client/components/wireframes/GroupsScreen/Group/LogDisplay/Message/index.tsx
