+++ client/components/wireframes/GroupsScreen/Group/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import { Header } from './Group/Header'
-import { Log } from './Group/Log'
+import { Header } from './Header'
+import { Log } from './LogDisplay'
 
 const MockMessages = [
   {
@@ -43,24 +43,13 @@ interface IgroupProps {
 
 const GroupScreen: React.FC<IgroupProps> = ({ group }) => {
   return (
-    <div className="">
+    <>
       <Header
-        name={group.name}
-        moderator={group.moderator}
+          name={group.name}
+          moderator={group.moderator}
       />
-
       <Log messages={MockMessages} />
-
-      <div className=" flex justify-center pt-12">
-        <button className=" relative font-thin text-xl text-green-500">
-          <div className="absolute inset-x-0 h-full -bottom-2 bg-gray-100 border border-gray-500 rounded-lg "></div>
-
-          <div className=" bg-blue-100 border border-gray-500 rounded-lg py-0.5 px-2 transition transform duration-200 hover:translate-y-2">
-            Enter Weekly Lottery Drawing
-          </div>
-        </button>
-      </div>
-    </div>
+    </>
   )
 }
 const AddGroupScreen: React.FC = () => {
