+++ client/tsconfig.json
@@ -15,6 +15,6 @@
     "jsx": "preserve",
     "incremental": true
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "components/wireframes/GroupsScreen/Group/LogDisplay/Index.tsx_"],
   "exclude": ["node_modules"]
 }
